Refactoring: move 'DummyTask' to utils so it can be reused.
